Make the data of DataExpressionSource public.
This is needed to be able to wrap this type of source in a new source
outside of metal and to avoid recreating the data.
Since the cache is mutable, we may want to improve this in the future
with a getCache method that returns a copy of the cache instead. Since
this may also cause performance problems, we made the field public for
now.