#392: Added a copy constructor to Ref.
This constructor is used to avoid converting an ImmutableList to an
array and then back to an ImmutableList again.
It also keeps the code simple.
Decided to deprecate two shorthands, to avoid confusion which shorthand
to use.
Adjusted tests to use the non-deprecated methods and then added tests to
ReferenceValueExpressionSemanticsTest to keep 100% test coverage.