#402: Define abstract class ImmuntableObject which caches the hash.
Yes, it is a mutable field for immutable objects, which is why it is
private and I created a separate class for it.
I have considered to calculate the hash within the constructor, but this
dropped performance as well, because then the hashes are always
calculated, even if you do not need it.