#402: Added TimeOut annotation again to performance test of ImmutableObject.
Fiddled the test numbers to find a scenario where the non-hash-caching
implementation would still finish within a measurable time. Added a
graph showing the different run numbers (on my computer) for comparison.
The TimeOut is set quite large in comparison to the hash-caching
implementation (2 seconds vs 55 ms), but that is to cover possible
GitHub hickups.

Processed review comments of @rdvdijk.