#405: Fix reading data from large concatenated values.
Previously, `values.head.slice().getData()` reads all the data in the
value and then copied an explicit part from it. In case you read small
blocks from a large value, this is very inefficient.

Added a getData variant to Slice class that allows specifiying the
offset next to the limit.