Moved ReversedImmutableList back as inner class of ImmutableList.
Since the ReversedImmutableList should only be retrieved by the reverse
method of the ImmutableList, it is also only created there. The
constructor can be made private this way.

Improved the number of times lists are copied when creating
ImmutableLists. Since we wrap the innerList with a unmodifiableList,
there is no need to make a copy of it first to avoid that the list is
accidentally changed by the ImmutableList implementation. Same for the
ReversedImmutableList. The list is now only copied within methods that
return a new list.

Added override implementation of additional methods within
ReversedImmutableList to make sure that every aspect of the object
returns a reversed view. This makes the implementation and test coverage
for ReversedImmutableList complete. The tests for the ImmutableList is
not yet complete.