Add a separate test for combiners when reducing streams.
These combiners are never executed for sequential streams, but since the
combiner should always be present, added a test to force a stream to run
parallel to trigger the combiner execution.

Decided to throw an exception in that case, because if the tokens are
parsed in parallel, it will mess up the parseState and it is unclear how
to combine two parseState into one. With an exception you can explain
why we stopped there.