#415: Move scopeDepth check to closeBranch.
The check within addBranch is too late. We can already detect the
illegal state when we want to close a branch such that we parse graph
gets into non-branched state. If the scopeDepth is then non-zero, which
should not be possible, then something was wrong before and we should
exit. This is a situation we can also test. Note that it will only occur
if the user creates a custom token that changes its scope delimiter
setting while processing. This is not the case in within Metal, but
found it valuable to add it anyway.