#415: Add ParseGraph to the list of classes to ignore for AutoEqualityTest.
The ParseGraph contain multiple private constructors and they are used
as copy constructors. It contains the cached value for scopeDepth that
matches the depth of the head. It is not possible to have two
ParseGraphs with the same head and tail, but with a different
ScopeDepth, but this is a case that will be testing using
AutoEqualityTest. We need to write our own equality test for parsegraph.