#415: Decided to test the ParseGraph by the AutoEqualityTest.
Although it is not a useful usecase, it is possible to create to
identical parse graphs with different scopedepth, by creating a token
that alternates its scope delimiter setting. That means that the scope
depth stored within the parse graph is actually not a cached value and
must therefor be included within the equals method. In that case, we can
test the equality of the ParseGraph by the AutoEqualityTest.